pyproject.toml
[tool.poetry]
name = "instagram-server-django"
version = "0.1.0"
description = ""
authors = ["selene-cho <selene.cho29@gmail.com>"]
readme = "README.md"
packages = [{include = "instagram_server_django"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
